% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin_fct.R
\name{burnin_fct}
\alias{burnin_fct}
\title{Burn-in function for algorithm requiring a step}
\usage{
burnin_fct(burnin, coef.burnin, scale = 1, format = "double")
}
\arguments{
\item{scale}{start value}

\item{format}{int or double}
}
\value{

}
\description{
Burn-in function for algorithm requiring a step
}
\examples{

u <- burnin_fct(10, 2)
plot(u)

v <- burnin_fct(10, 3/4, 0.5)
plot(v)
}
