% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vector_ref.R
\docType{class}
\name{fct_vector}
\alias{fct_vector}
\title{Vector of function}
\description{
The purpose of this object is to provide a new type of numeric array whose components are calculated if and only their access was requested.
So for example if one of the components is costly and rarely called this will not impact the rest of the calculations.
}
\section{Fields}{

\describe{
\item{\code{fct}}{list. Several name or unnamed functions that compose the vector.  Warning they must have all the same argument. They must return a numeric object, no length limited}

\item{\code{dimention}}{list. The length of the numeric return of each function. Default value assume that of function return a one dimension numeric.}
}}


\examples{
require(SAEM4JLS)
vec <- fct_vector(function(x,y) x*y,
                 function(x,y) x+y,
                 function(x,y) c(2*x, 2*y), dim = c(1,1, 2))
vec$eval(1, 2, i = c(1,2))
vec[1,2](1,2)
#Warning: the vector’s extract operators [ do not return a fct_vector but the eval function of the extracted vector
#If you want the extracted vector use the list’s extract operator [[
G <- vec[1,2] # is a function
G(1,2)

H <- vec[[1,2]] # is a fct_vector
H$eval(1,2)
}
