% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vector_ref.R
\name{get_dimension_index}
\alias{get_dimension_index}
\title{Return the good index for each component of a fct_vector}
\usage{
get_dimension_index(dim)
}
\description{
Return the good index for each component of a fct_vector
}
