% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGD.R
\name{grad}
\alias{grad}
\title{Calculation of the averaged gradient}
\usage{
grad(grad.fi, theta, mini.batch, normalized.grad, args = list())
}
\arguments{
\item{mini.batch}{support of the sum}

\item{args}{additional argument}
}
\description{
Compute grad(f)(theta) = sum(grad(f_i)(theta))
where the support of the sum can be chosen
}
