% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH.R
\name{MH}
\alias{MH}
\title{Metropolis Hastings MCMC}
\usage{
MH(niter, x, transition, loglik, ..., verbatim = F)
}
\arguments{
\item{niter}{number of iteration}

\item{x}{the initial value of the MCMC}

\item{transition}{function that apply the transition between x and the new proposal}

\item{loglik}{function that return the acceptation rate, the first argument must be the proposition value in the algorithm}

\item{...}{the parameter for loglik function}
}
\description{
Metropolis Hastings MCMC
}
