% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_burnin.R
\name{MH_burnin}
\alias{MH_burnin}
\title{Burn-in with Metropolis Hastings}
\usage{
MH_burnin(x, n)
}
\arguments{
\item{x}{object return by a MH function}

\item{n}{number of samples being discarded}
}
\value{
the MCMC without the n first samples
}
\description{
Burn-in with Metropolis Hastings
}
