% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_High_Dim.R
\name{MH_High_Dim}
\alias{MH_High_Dim}
\title{Metropolis Hastings MCMC}
\usage{
MH_High_Dim(niter, x, transition, loglik, ..., verbatim = F)
}
\arguments{
\item{niter}{number of iteration}

\item{x}{the initial value of the MCMC}

\item{transition}{function that apply the transition between x and the new proposal}

\item{loglik}{function that return the acceptation rate, the first argument must be the proposition value in the algorithm}

\item{...}{the parameter for loglik function}

\item{verbatim}{boolean value, true will return the whole chain, false will return by default only the last value}
}
\description{
Will run the metropolis hasting algorithm on the rows of x.
For example if we have x = (x_{i,j}), we will do an iteration of MH for each x_{i,.}
}
