% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_High_Dim_future.R
\name{MH_High_Dim_future}
\alias{MH_High_Dim_future}
\title{Metropolis Hastings MCMC}
\usage{
MH_High_Dim_future(niter, x0, sd, loglik, ..., verbatim = F, cores = 1)
}
\arguments{
\item{niter}{number of iteration}

\item{x0}{}

\item{loglik}{function that return the acceptation rate, the first argument must be the proposition value in the algorithm}

\item{...}{the parameter for loglik function}

\item{verbatim}{boolean value, true will return the whole chain, false will return by default only the last value}
}
\description{
Will run the metropolis hasting algorithm on the rows of x.
For example if we have x = (x_{i,j}), we will do an iteration of MH for each x_{i,.}
}
\details{
!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!
This code is paralized! Your variables must be arranged in columns and the individuals must be on the rows!
Otherwise the first variable to be calculated by the algorithm will be forced to take absurd values to counterbalance the absence of update of the other variables!
Keep in mind that the variables must be able to balance each other and this is possible IF AND ONLY IF they are on the columns!
on this, good programming and live long#'
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
}
