% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_res.R
\docType{class}
\name{MH_res-class}
\alias{MH_res-class}
\alias{MH_res}
\title{Metropolis Hastings Object}
\value{

}
\description{
Metropolis Hastings Object
}
\section{Fields}{

\describe{
\item{\code{value}}{matrix.}

\item{\code{acceptation}}{data.frame.}

\item{\code{chain}}{data.frame.}
}}

