% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLME_obs.R
\name{NLME_obs}
\alias{NLME_obs}
\title{Non Linear Mixed Effect observation}
\usage{
NLME_obs(G, ng, t, param, NLME_fct)
}
\arguments{
\item{G}{number of group}

\item{ng}{number of individuals per group}

\item{t}{vector containing the times the observation}

\item{param}{list of parameter( rho2, sigma2, mu, omega2)}

\item{NLME_fct}{an non linear function (t, phi, eta)}
}
\value{

}
\description{
Non Linear Mixed Effect observation
}
