% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_High_Dim_plot.R
\name{plot.MH_res}
\alias{plot.MH_res}
\title{plot resultat of MH function}
\usage{
\method{plot}{MH_res}(x, dim = 1, value = T, acceptation = T)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plot resultat of MH function
}
