% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEM_plot.R
\name{plot.SAEM}
\alias{plot.SAEM}
\title{plot the result of the SAEM funciton}
\usage{
\method{plot}{SAEM}(res, MCMC = T)
}
\arguments{
\item{MCMC}{}
}
\value{

}
\description{
plot the result of the SAEM funciton
}
