% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGD.R
\name{prox}
\alias{prox}
\title{proximal operator for a vector}
\usage{
prox(theta, gamma, alpha, lambda)
}
\arguments{
\item{theta}{vector argument}

\item{gamma}{step}

\item{alpha}{choice between ridge and lasso}

\item{lambda}{tuning parameter of elastic net}
}
\description{
proximal operator for a vector
}
