% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGD.R
\name{proxi}
\alias{proxi}
\title{Proximal operator for one component}
\usage{
proxi(theta, gamma, alpha, lambda)
}
\arguments{
\item{theta}{scalar argument}

\item{gamma}{step}

\item{alpha}{choice between ridge and lasso}

\item{lambda}{tuning parameter of elastic net}
}
\description{
Proximal operator for one component
}
