% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEM.R
\name{SAEM}
\alias{SAEM}
\title{Stochastic Approximation of Expectation-Maximization}
\usage{
SAEM(
  niter,
  niter.MH,
  u,
  param,
  Phi,
  exhaustive,
  Z,
  simulation,
  maximisation,
  eps = 0.001,
  verbatim = F
)
}
\arguments{
\item{simulation}{function for simulation step}

\item{maximisation}{function for maximisation step}

\item{M}{number of MCMC to compute with MH}
}
\value{

}
\description{
Stochastic Approximation of Expectation-Maximization
}
