% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin_fct.R
\name{burnin_fct}
\alias{burnin_fct}
\title{Burn-in function for algorithm requiring a step}
\usage{
burnin_fct(burnin, coef.burnin, scale = 1)
}
\arguments{
\item{scale}{}
}
\description{
Burn-in function for algorithm requiring a step
}
\examples{

u <- burnin_fct(10, 0.1)
plot(1:20, u(1:20))

v <- burnin_fct(10, 3/4, 0.5)
plot(1:20, u(1:20))
plot(1:20, v(1:20))
}
