% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_plot.R
\name{plot.MH_res}
\alias{plot.MH_res}
\title{plot resultat of MH function}
\usage{
\method{plot}{MH_res}(x, name, nrow, ncol, var = c("chain", "acceptation"))
}
\arguments{
\item{...}{}
}
\description{
plot resultat of MH function
}
