% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEM_plot.R
\name{plot.SAEM_res}
\alias{plot.SAEM_res}
\title{plot the result of the SAEM funciton}
\usage{
\method{plot}{SAEM_res}(res, var = c("parameter", "MCMC", "acceptation"), true.value)
}
\arguments{
\item{res}{'SAEM' type response variable}

\item{var}{what to plot}

\item{true.value}{the true value of the parameter, if it is not missing the plot function will return a knitr array summarizing the estimates with RMSE calculation}
}
\value{

}
\description{
plot the result of the SAEM funciton
}
