% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEM.R
\name{SAEM}
\alias{SAEM}
\title{Stochastic Approximation Expectation Maximization (SAEM) algorithm}
\usage{
SAEM(
  niter,
  sim.iter,
  parameter0,
  var0,
  Phi,
  exhaustive,
  maximisation,
  simulation,
  adptative.sd = NULL,
  burnin = NULL,
  coef.burnin = 1,
  eps = NULL,
  verbatim = F,
  RDS = NULL
)
}
\arguments{
\item{niter}{number of iterations for SAEM loop}

\item{sim.iter}{number of iterations for Gibbs sampler algorithm}

\item{var0}{list of latent variables (with var names) MUST BE VECTOR}

\item{maximisation}{function for maximisation step}

\item{simulation}{function for simulation step}

\item{adptative.sd}{target sd value}

\item{burnin}{step at which burn-in period ends}

\item{coef.burnin}{1/(step+1-burnin)^coef.burnin}

\item{eps}{Stop criterion}

\item{verbatim}{1 = message only, 2 add MCMC, 3 add full MCMC}
}
\value{

}
\description{
SAEM algorithm perform parameter estimation for nonlinear mixed effects
}
