% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPGD.R
\name{SPGD}
\alias{SPGD}
\title{Stochastic Proximal Gradient Descent}
\usage{
SPGD(niter, theta0, step = 1e-06, grad.fi, n, f, ..., verbatim = F)
}
\arguments{
\item{niter}{number iteration of the gradient descent}

\item{theta0}{init point}

\item{step}{of the gradient descent}

\item{grad.fi}{function that return grad(f_i, ...)}

\item{f}{function that return f}

\item{...}{additional argument for the grad function (and f)}

\item{m}{number of iteration for the average computation of theta}
}
\description{
compute a minimization of : f(theta) + pen(theta) where pen is the elastic net
}
